+++ labbook_BE/app/models/DB.py
@@ -6,6 +6,7 @@ import re
 
 from flask import current_app
 from app.exception.DBException import DBException
+from app.models.Constants import Constants
 from app.models.Logs import Logs