+++ labbook_BE/app/models/Product.py
@@ -2,7 +2,7 @@
 import logging
 import mysql.connector
 
-from app.models.Constants import *
+from app.models.Constants import Constants
 from app.models.DB import DB
 from app.models.Logs import Logs