+++ labbook_BE/app/services/GeneralRest.py
@@ -99,7 +99,9 @@ class InitVersion(Resource):
         ini = Various.getLastInitVersion()
 
         if ini['ini_stat'] == 'Y':
-            if Various.updateTranslationsTable('en_GB'):
+            locales = ['en_GB', 'pt']  # add Portuguese
+            ok = all(Various.updateTranslationsTable(loc) for loc in locales)
+            if ok:
                 Various.updateInitVersion(ini['ini_ser'], 'N')
 
         self.log.info(Logs.fileline() + ' : TRACE InitVersion ini_ser=' + str(ini['ini_ser']))